Refactored Property parsing
Now properties are not parsed, but instead they are validated. Instead
of receiving an Object and parsing it to a PropertyValue, a PropertyValue
is received and gets validated.
Static methods have been added to PropertyValue to make easier to create
these objects.
The next step is to delete PropertyValueParser and do that validation
in the property. Then, update the javadoc.

+++ pgconfig/src/main/java/com/ongres/pgdeploy/pgconfig/properties/DataType.java
@@ -29,7 +29,7 @@ package com.ongres.pgdeploy.pgconfig.properties;
  */
 public enum DataType {
   STRING(String.class),
-  INTEGER(Integer.class),
+  INTEGER(Long.class),
   DOUBLE(Double.class),
   BOOLEAN(Boolean.class);
